Implement build of TSEM loadable models.
The V4 release of TSEM has support for the implementation of
workload specific models using kernel loadable modules.  This
commit provides an example implementation of a TSEM kernel
loadable module.

There are three ways of modifying the processing of a system
event:

- bypasses

- event initialization

- event mapping

The bypasses array defines whether or not an event is to be
processed.  Modules can select which events they want to handle
by specifying 'true' values for an events slot in the bypasses
array.

A module can specify how to handle the initialization of the TSEM
event description by specifying a workload specific method to be
called by the in-kernel based event initialization
infrastructure.

Finally, a module can specify how an event is to be mapped into a
security event coefficient by providing a model specific mapping
method.

The sample module demonstrates how the equivalent of the Linux
Integrity Measurement Architecture (IMA) can be implemented via a
TSEM loadable model.

The only events processed are file opens and memory mapping of
files.  The event initialization code only selects for events
that have or are capable of root privilege levels.  The
initialization code selects only events that involve the opening
of files with executable status.  In the case of memory mapping,
only events that involve the manipulation of the executable
status of the memory are selected.

The mapping method that is provided returns, as the security
coefficient, the checksum of either the file or the mapped memory
region.

The build of the contents of the TSEM directory is triggered by
the presence of the following build configuration variable:

BUILD_KERNEL_SOURCE

If this variable is not set the TSEM directory will be bypassed
by the build process.  If set, the contents of this variable will
be used as the kernel source context for building the kernel
module.