Simplify generation of sorted coefficients.
The TSEM->get_state method was previously using a sorted array of
SecurityPoint objects in order to generate the model security
state.  The update to the Gaggle object created issues with this
approach since the Gaggle->get method was returning a pointer to
an object rather than the object itself.

This commit implements an alternate and simpler scheme that
creates a Buffer object of the current security coefficients and
then sorts this Buffer object.

This has the disadvantage of using more memory but it immediately
addresses the segmentation faults that were occuring with user
based modeling.