Update TSEM documentation.
The in kernel tsem.rst file was updated to include the 'model'
argument that can be passed to namespace creation to indicate the
name of a TSEM kernel loadable model that should be used for the
security modeling namespace.

In addition the kernel documentation was updated to include the
'lock' directive that is used to disallow any further changes to
the set of available models.

Update the Quixote version of this document to reflect these
changes.