Drop privileges in quixote-export utility.
One of the features of TSEM's export only mode is that security
reconaissance can be conducted with the need for elevated
privileges or invasive taps or probes.  In order to realize this
more freely this commit changes drops privileged credentials in
favor of an alternate non-privileged set of credentials for the
processes that are monitoring the child process and exporting
events.

The -s command-line argument was selected to specify the user
identity that process privileges are to be changed to.  By
default the 'nobody' user is used.

The user argument was added to the fire_cartridge() function in
order to pass the desired user down into the infrastructure that
initiates the proceses that control and initiate the security
modeling namespace.

The user variable was already used to specify the user identity
for the MQTT broker so that variable was change to tsem_user.

A new helper function named _set_user() was added to assist the
fire_cartridge() function in dropping its privileged status.

The privileges for the launch process are dropped just before the
invocation of the child_monitor() function.  The privileges for
the monitor process are dropped after the fork that supports the
launch of the desired workload.

Additional work on this infrastructure is needed since the
control proceses are running at reduced privilege levels and
would therefore have issues with potentially acting on the
subordinate processes.  That work will be left to subsequent
updates to this infrastructure.