Move use of TSEMcontrol object into namespace creation function.
Work is being done to review what resources are kept open across
the process forks that are needed to create the trust
orchestration architecture.

An initial review and experimentation indicates that the use of
the TSEMcontrol object could be isolated exclusively to the
setup_namespace() function.

The initialization of the previous static TSEMcontrol
implementation was removed in favor of isolating the
functionality to the setup_namespace() function.