Prevent leakage of file descriptors in subordinate processes.
Testing with the quixote-export utility indicated that the
monitor process and the workload process were inheriting file
desriptors from their parent processes that were not needed
by the subordinate processes.

The following four types of resources were being leaked into the
child processes:

- Event export pseudo-file.

- Event output file if file based capture was being used.

- Debug output file.

- Event pipe descriptors.

In the case of MQTT broker output the socket connections were
also being inherited by the workload process.

This commit addresses the first four types of inappropriate
inheritance.

The call that opens the file used to export the event
descriptions was modified to take the O_CLOEXEC option which
causes the file to be closed when the workload process is
executed.

A test is made for the presence of a valid output file after the
namespace creation process is created.  If an output file is
found the file object is destroyed in order to close the file in
the process.

In the process that runs the workload an fclose call was added to
the clause that was testing for whether or not debugging was
enabled.

In this process a close of the write side of the event
transmission pipe was also added in order to prevent the pipe
descriptor from being conveyed into the workload process.

These changes do not address the MQTT socket connections from
being conveyed into the workload process.  A simple attempt to
address this in the same manner as the output file generates
issues with the signals being generated to control the broker
client threads.

Addressing this will require a more structural change in the code
that will be forthcoming in subsequent commits.