Fix looping in namespace monitor.
Testing with the updated quixote-export implementation indicated
that the polling loop in the child_monitor() was running in a
continue loop.

This was detected when debugging was enabled and the debug file
grew large very rapidly with no indication of events being
processed.

The cause of the problem was a long standing issue with the
configuration of the poll() function.  The quixote-export utility
only has to listen on a single file descriptor rather than two
file descriptors, as is the case then the child_monitor()
function is monitoring for orchestrator commands as well as
events from the workload.

This is a long standing issue that was surfaced by the
reformulation of the code.

Addressing the problem simply consisted of specifying the number
of file descriptors that are being monitored.