Implement TSEMworkload object.
Up to this point there has been a lot of code duplication across
the various userspace trust orchestrators.  This commit
introduces the TSEMworkload object to consolidate the management
of a TSEM workload into a single object.

The two primary methods from a functional basis are as follows:

TSEMworkload->run_monitor

TSEMworkload->run_workload

In a userspace TSEM trust orchestrator there are typically three
processes:

1.) Security monitor process.

2.) Namespace monitor process.

3.) Workload process.

The Security Monitor (SM) process is the initial process that
starts a TSEM workload.  The responsibility of this process is to
field the security events from the workload and either handle
their export or their processing by a Sancho security model.

The SM process forks to create the Namespace Monitor (NM)
process.  This process creates a new TSEM security modeling
namespace and forks again to generate the workload process.

Both the NM process and the workload process will run in the
isolated security modeling namespace.  The NM process is
responsible for reading the security events generated by the
workload process and forwarding them through a pipe to the SM
process.

By default a TSEMworkload process will run in shell mode where
the currently defined SHELL process is run in an isolated
security modeling namsepace.

The TSEMworkload->set_execute_mode method is used to specify an
argument count and list that describes an executable to be run in
an isolated security modeling namespace.

The TSEMworkload->set_container_mode is used to specify the name
of a runc bundle directory and container name that is to be run
in an isolated security modeling namespace.

The TSEMworkload->configure_export method is used to define the
parameters for a security modeling namespace that is to export
its events.

The TSEMworkload->set_debug method is used to specify a File
descriptor that debug messages are to be sent to.

This is a preliminary implementation of the object and there are
expected to be multiple changes as additional trust orchestrators
are updated to use this object for the management of a workload
in a security modeling namespace.