Prevent segmentation faults in non-debug invocations.
Closing of the debug file was not being protected by a test for
whether or not debugging was enabled in the TSEMworkload object.

This resulted in the fclose() function being called with a NULL
pointer when debugging was not enabled.

Each instance of where the fclose() function was being called is
immediately after the final debug statement that is printing what
is being executed.  Place the print code and the fclose call in a
clause that is protected by the conditional checking for whether
or not debugging is enabled.