Fix broker output for quixote-export.
Testing with the updated quixote-export utility indicated that
broker based output was not working properly.  This commit
reflects the changes needed restore output to an MQTT broker.

The root problem with broker output was that the broker
connection was not being made due to the code reorganization that
has been conducted.

In order to avoid code duplication the following two functions
were introduced:

open_file()

open_broker()

The code for opening an output file was moved into the
open_file() function so that the export_root() function and the
run_workload() functions use the same code for opening the output
file.

The same strategy was followed for opening the MQTT broker
connection.