Fix repeated output in quixote-export.
This commit addresses the generation of incorrect output when the
quixote-export utility was running in root export mode.

The quixote-export utility appeared to be not generating output
when the events were tracked with an MQTT broker client.  Events
were being generated but the same sequence of events were always
being output, a fact that appeared to make the utility look like
it was not generating output.

The root of the problem was traced to the Gaggle->rewind_cursor
method call at the top of the _export_events() function.  This
was causing the queued events to be over-written rather than the
events being added to the end of the queue.

The incrementing of the queued event count was transferred to the
_get_event() function since that is the location that actually
increases the number of events that are queued.  At the end of
this function the queued event count is cleared to zero and the
cursor is reset on the Gaggle output object.

A call to the _output_events() function was added immediately
after the events that were queued in the kernel handled and
before entering follow mode if follow mode is enabled.