Always issue an lseek call after every read of root export file.
In order to maintain consistency with the code in the
TSEMworklaod object, the lseek command in the _get_events()
function was moved to after the done: label.  This causes any
read of the root event file to be followed by an lseek() function
call to reset the pseudo-file to the first record in the file.