Use the type of namespace configured for the object.
At the time the TSEMworkload object was created the type of the
namespace was configured to be of a type of export, given that
code was copied intact from the quixote-export.c code.

Change the type specification to use the type of namespace that
was configured for the object, using the TSEMworkload->type
variable.