Update TSEMworkload object to support externally modeled namespaces.
This commit modifies the TSEMworkload object so that it will
support running a workload that is externally modeled.

One of the significant changes was to implement a TSEMcontrol
object in the state information of the TSEMworkload object.  The
control object is opened by the TSEMworkload object initializer.
This allows the workload control file descriptors to be managed
by the TSEMworkload object based on whether or not the security
monitor or the workload processes are being run.

The following three additional methods were added as well:

TSEMworkload->configure_external

TSEMworkload->release

TSEMworkload->discipline

The ->configure_external method was added to allow an application
to specify that the workload object was to be configured for
running a TSEM workload that is externally modeled.

The ->release and ->discipline methods were added as inheritance
wrappers around the TSEMcontrol->release and
TSEMcontrol->discipline methods.  This is needed so that the
function that is called to handle workspace events has the
ability to release a process as either being trusted or
untrusted, depending on the results of the model.

The TSEMworkload->run_monitor method was updated to include code
that supports the polling of the file descriptor for the
management socket attached to the trust orchestrator.

The TSEMworkload->run_monitor method was updated to take an
additional argument that specifies that function that is to be
called to handle control plane requests to the trust
orchestrator.  The function implementing the command handler
takes as arguments the object describing the command socket and a
Buffer object that will contain the command that is read over the
socket.

The O_CLOEXEC option was removed from the open() call for the
namespace event pseudo-file since it was causing a the userspace
trust orchestrator process to hang.  The functionality of the
O_CLOEXEC was replaced by simply closing the file descriptor for
the event file in the context of the workload process.