Update quixote-export for changes in the TSEMworkload object.
The updates to the TSEMworkload object needed to support an
externally modeled workload required minor changes to the
quixote-export utility.  The most principal change was to add a
NULL argument to the TSEMworkload->run_monitor calls to indicate
that no control plane socket was available.

The other changes were to drop the definition of objects that are
no longer needed at static scope since their functionality has
been incorporated into the TSEMworklaod object.