Implement current namespace designator as a functional local variable.
The Current_Namespace static variable was used to indicate that
the current user namespace, rather than the initial namespace,
should be used to resolve the numeric value of discretionary
access control identities.

The conversion of quixote-us to using the TSEMworkload object
allows this boolean variable to be used in function local scope
rather than static scope.

Drop the static declaration of the variable in change for a
version declare in the main() function.  Also changes the
capitalization to the conventions used for a function scoped
variable.