Add support for files created in a security modeling namespace.
The TSEM driver has support for denoting files that have been
created in the context of security modeling namespace.  It does
so by associating an instance number of the file that is created
and the task identity of the process that created the file.

In the case of a temporary file the following measurement series
is used:

PATHNAME_LENGTH || PATHNAME || INSTANCE || OWNER

Where PATHNAME is the path description up to the directory that
contains the temporary file.

The Cell object was updated to detect and parse the instance
definitions for a path.  The output of the Cell->format method
was updated to reproduce the instance information that was in the
original event description.

Adding the measurement stream for a temporary file was segragated
into an independent helper function named _add_temp_path().

The following additional general help function was added to
encapsulate the output of a String ojbect into a measurement
buffer:

_add_String

For parsing an event description the following helper function
was introduced:

_get_u64

Which extracts a 64 bit value from a JSON key and converts it
into a 64-bit unsigned numeric value.

With this update the quixote-us utility is now capable of
generating invariant state values from a runc container booting
into a shell.