Fix parsing of task_kill event.
An upgrade to version 1.2.2 of the runc utility resulted in
externally modeled namespaces generating errors when task_kill
events were generated.

It appears that the parsing of these events had not been updated
when the kernel implementation of the JSON encoding of the event
was changed.

The following key name changes were implemented in the processing
of the event:

cross -> cross_ns

signal -> sig

The reformatting of the event information was updated so that the
reformatted event matched the encoding of the event that was
generated by the kernel.