Update TSEM integrity model to match kernel implementation.
The integrity model was updated to match changes that were
implemented in the TSEM kernel implementation.

The kernel now implements an architecture where a model specifies
the events that are to be processed, rather than the events that
are to be bypassed.

The array containing the events to be processed was renamed from
bypasses to event_handlers.  The TSEM_FILE_OPEN and
TSEM_FILE_MMAP events were registered as events to be handled.

The method names for the tsem_context_ops structure were renamed
to match the member naming implementation that is now used in the
kernel.  For the purposes of this model the method that handles
the mapping of these two events was changed from 'init' to
'cell_init'.

Since named array initializations are now being used the code
that initialized the events to be handled in the
initegrity_init() function can now be dropped.