Implement utility for comparing execution trajectories.
A common need in the development of TSEM security models is to
compare two executions of a workload to verify that they reach
identical security states.  If not it is useful to development to
understand the difference between the two execution trajectories.

In order to support this functionality this commit implements the
following utility:

diff-states

This utility has three separate modes:

-A (auto):	Prompt for the execution of two workloads and compares
		the trajectory of the two workloads.

-C (compare): 	Compare/check the results of two workload
		executions.

-G (grab):	Grab the state, security coefficients and
		trajectory of a security modeling namespace.

This utility supports a workload that is run either as a process
heirarchy or in a runc container.  The -p and -c command-line
arguments specify that the workload was executed as either a
process or container based workload.

The argument to this function is the name of the workload that
is being executed.

The 'grab' mode also requires the specification of a version
identifier for a specific run of the workload.  The compare
functionality assumes that the version numbers will be either 1
or 2.

At the current time the utility takes a rather simplistic
approach to comparing the execution trajectories of two
workloads.

If the state values match the utility indicates the trajectories
are equal.

If a state mismatch occurs the utility prints out the first event
that is different and the two event descriptions.