Implement the pathname reference.
The kernel TSEM implementation was updated to implement the
export of the reference that was used for a pathname.

The following string values are exported:

root
chroot
namespace

Where the root string indicates that the pathname is relative to
the filesystem root.

The chroot string indicates that the process generating the
pathname was in a chrooted environment.

The namespace string indicates that the pathname was generated in
a mount namespace.

The implementation of this new field was straight forward.  A
String object was added to the path structure used in the CELL
definition.  The value of the type key in the JSON path
description was parsed and the string valued was used to populate
the String object.

The contents of the String object is used to extend the
pre-image buffer that from which the CELL identity is computed.

The _format_file() and _dump_file() functions were updated to add
the new key type to the event description and the dump output for
the object.