Drop reference to replacement of IMA.
A review comment from Paul Moore was that the section of the
documentation that was referring to the ability of TSEM to
implement any security model was an allusion to the notion that
TSEM could replace IMA.

Reword the documentation to reflect the fact that TSEM is capable
of implementing a basic file integrity model.  Also include a
pointer to the fact that the Quixote userspace tools have an
implementation of such a model as a TSEM loadable module.