Clarify inspitation for security event modeling.
In review comments Casey Schaufler indicated that it was
unreasonable to assume that kernel developers would have an
understanding of the Turing Abstract Machine Model.

Clarify the documentation to include a link to a Wikipedia
article discussing Turing machines.

In addition, the documentation now explicitly calls out the
relationship between the COE in security event modeling and the
'head' in the Turing model of computation.  Also explain that the
CELL is a reference to the unique security measurement points in
an execution trajectory.