Implement direct process identification.
The TSEM kernel driver was updated to remove the process
authentication system that had been initially implemented in
order to address complaints that TSEM would be vulnerable to a
process substitution attack.

To replace the process authentication system a unique task
identifier was implemented based on a 64-bit integer counter that
can be guaranteed to never overflow.  This counter is incremented
each time a new task is created.

For each security event that is exported to userspace the task
identifier value is exported along with the PID value for the
task.  The kernel expects commands of the following form through
the TSEM control plane to release a process:

trusted PID TNUM

untrusted PID TNUM

Where the PID and TNUM values are those that were exported with
the event.

The TSEMworkload->displine and TSEMworkload->release methods were
updated to take the task number as an additional argument that
will be transmitted through the control plane.

The TSEM model object was updated so that the
TSEM->discipline_pid method was passed a pointer to a variable
that will be loaded with the task number of the event.

The SecurityEvent->get_pid method was updated to include a
pointer to a variable that will be loaded with the task number of
the event.

Support for passing the key value through the control plane was
removed since it is no longer needed.

As of this update only the userspace trust orchestrator was
updated to use the new direct process identification.  Other
trust orchestrators will be updated in future updates.