Reset the path reference type.
Testing of event processing demonstrated a regression where the
the 'type' key was being duplicated when the JSON encoding of the
event string was reproduced.

The following is an example of the replicated key/value:

"type": "namespacenamespace"

This was secondary to the fact that the Cell->reset method was
failing to reset the String object that holds the type
description.

Add the reset of the String object to the reset method corrected
this issue.