Handle escaped pathnames.
The TSEM kernel driver was updated to 'escape' double quotes that
are in pathnames.  This is to prevent a filename with quotes from
breaking the parsing of an event description.

The _get_key private function in the TSEMparser object
implementation was updated to convert the escaped string back to
a normal string with embedded quotes.

The _escape_path() private function was added to the Cell object
in order to properly escape a pathname since the TSEMparser
object will convert the escaped string back to a string
containing the initial quotes.

The test-event utility was updated to to have a pathname that
contained a double quote in order to test this parsing
functionality.