Open TSEM control plane before loading model.
A regression was noted with the kernel based trust orchestrator
generating a segmentation fault when a security model was loaded.

This was caused by the TSEMcontrol object not being initialized
before the the model was loaded.  The TSEMcontrol object is used
to write the security state coefficients into the kernel and
without being initialized the object was NULL, thus generating a
segmentation fault when the TSEMcontrol->add_state method was
invoked.

This was fixed by moving the TSEMcontrol object initialization to
immediately before the loading of the security model.