Do not close the control plane file prematurely.
After the segmentation fault in the kernel based trust
orchestrator was fixed the orchestrator was still generating a
segmentation fault.

This was caused by the fact that the File object that was being
used to access the control plane file was being closed at the end
of the TSEMcontrol->control_ns method.  This was no doubt
inadvertent since the TSEMcontrol->reset method also has a
destructor for that File object.

Remove the destructor from the TSEMcontrol->control_ns method.