Initialize TSEM model before loading security model.
In a situation similar to what was corrected with the in kernel
orchestrator the initialize of the TSEM object was moved to
before the security model was opened and loaded.

Referencing the method to load security state coefficients into
the model was causing a segmentation fault due to the
uninitialized object.