Prevent segmentation fault caused by improper format specifier.
After the segmentation fault was corrected by proper
initialization of the model (TSEM) object the userspace trust
orchestrator was still generating a segmentation fault.

This was traced to the fact that an improper format specifier was
being used in the String->add_sprintf method call that was being
used to generate the TSEM control plane call in the
TSEMcontrol->discipline method.

A string (%s) format specifier was being used for the tnum value
rather than the (%lu) format specifier that was appropriate for a
64-bit integer value.

Correct the format specifier to prevent the segmentation fault.