Add task number to debugging messages.
In debugging the kernel regression that was causing exported
security event descriptions to have a task number of zero it was
deemed of use to have the debug messages that define the task to
be disciplined to contain the task numbers.

Add the task number of the event to the debug messages that are
logged.