Update handling of writes to control plane.
Improvements to the userspace trust orchestrator required updates
to the TSEMworkload object that in turn required updates to the
TSEMControl object that issues commands to the TSEM control
plane.

When an asynchronous shutdown of a workload is executed there is
the chance that a process that is having its trust status set
will no longer exist.  This is a normal situation but will cause
writes to the control plane to return an error value.

Modify the ->discipline method and the _write_cmd private
function so that they return an error status but do not issue
error messages as to the location of the fault.

The previous implementation of the _write_cmd was allocating a
Buffer object in order to use the File->write_Buffer method.  The
HurdLib library had been updated to provide a File->write_String
method so that method was used in order to eliminate the object
allocation and destruction.