Update TSEMworkload object to support workload shutdown.
When an externally modeled workload, running in enforcing mode,
experiences a security fault from an event, running in atomic
context, the only alternative is to terminate the workload.

The ability to shutdown the workload was lost when the
TSEMworkload object was implemented.  This commit implements the
infrastructure needed to execute the shutdown of a workload.

The former kill_cartridge() function was replaced with the
TSEMworkload->discipline method.  The functionality of the
function was largely maintained.

If the workload is non runc based a SIGHUP command is transmitted
to the workload process.

If the workload is runc based the 'runc kill' command is used to
send a SIGKILL signal to the init process of the workload.

An additional check for the presence of a SIGCHLD signal was
added to the ->run_monitor method in order to make sure that the
death of a child process was immediately captured when multiple
events were being read processed from a single read of the
control plane file that is delivering workload event
descriptions.