Implement externally modeled workload shutdown.
When the external trust orchestrators were converted to use the
TSEMworkload object, the ability to shutdown a workload that
experienced a faulty in an event that was running in asynchronous
context was lost.

The TSEMworkload object had been previously updated to include a
->shutdown method that implements the shutdown of the workload.
This commit uses this infrastructure to execute workload shutdown
when a security fault was detected in an event running in atomic
context when the workload was running in enforcing mode.

The following statically scoped variable was added in order to
provide an indication that the workload was being shutdown:

In_Shutdown

This boolean variable is set to a 'true' value when a sealed
workload experiences a violation in enforcing mode.

The trust status of the process generating the event is set to
untrusted that will cause any further events that the process
generates to return an error value.

Given that the workload may be in an unstable state while it is
undergoing shutdown subsequent calls to release or discipline a
process are not checked for an error return so that the trust
orchestrator is not prematurely terminated until the workload
process exits.