Implement enforced modeling for external orchestrators.
The userspace trust orchestrator was not able to place a workload
in enforcing mode since the kernel modeling context was not
update to change the event actions from logging to enforcement.

In order to implement this generically the TSEMworkload object
was updated to accept an additional boolean argument to the
TSEMworkload->configure_external method.  The value of this
argument is placed in the state information of the object.

The _setup_namespace() private function was updated to test the
value of this variable and issue the TSEMworkload->enforce method
if enforcing mode has been enabled.

The quixote-us orchestrator was setting two separate variables
when the -e option was encountered on the command-line.  The
functional local variable was dropped in favor of the global
variable that is sampled elsewhere in the implementation.

This new architecture allowed the enforce variable that was being
passed to the run_workload() function to be dropped.