Improve handling of process based externally modeled workloads.
Workloads that were being modeled by the quixote-us trust
orchestrator were exhibiting poor shutdown behavior, particularly
when triggered by a violation in a security event that was
running in atomic context.

In order to better handle the shutdowns an extra parameter was
added to the TSEMworkload->shutdown method.  This parameter
allows the caller to specify the signal that is to be used for
the shutdown.

The quixote-us.c:add_async_event() function was updated to send
the SIGTERM signal to the namespace monitor process, that process
in turn sends the appropriate signal to the workload process.

The shutdown calls in the TSEMworkload implementation itself were
modified to use the new ->discipline method call syntax and to
specify the signal that was appropriate for the position in the
process stack from which the ->shutdown method was being invoked.

This patch also implements a policy of clearing the signal flag
that was detected as being set when the signal was detected.