Check for remote closure before sending data.
A regression was noted when a control command was issued to the
quixote-us external trust orchestrator.  If the output of the
command was piped into another program and prematurely terminated
the monitor process immediately terminated.

The issue was tracked down to the fact that the trust
orchestrator continued to try and send output over the UNIX
domain socket after the quixote-console program had terminated.

The issue was addressed by implementing the _check_write()
function that issues a poll command against the I/O file
descriptor of the socket.  The file desciptor is polled for the
POLLHUP and POLLOUT events which represent either the remote end
of the socket being closed or the fact that the socket is ready
to receive a write.

If the remote end closes the pipe the eof state variable for the
LocalDoct instance is set to true and the write command returns
with a true value.

A poll error causes a false value to be returned for the write.