Properly handle command socket closures.
The LocalDuct object was updated to check the remote end of the
UNIX domain socket before attempting a write to the socket.  This
commit updates the TSEMworkload object to properly handle the
closing of the socket.

The previous implementation of the TSEMworkload->run_monitor was
substituting the I/O file descriptor of the socket for the socket
descriptor in the poll command event descriptions.  This caused
the closure of the remote end of the socket to be only detected
in the primary event poll loop.  The implementation also violated
the concept of least surprise.

This new implementation expands the size of the poll event
descriptions to three array members when a connection is accepted
on the control socket.  The third event description is for
polling for input from the control socket.

When input is detected the command handler for the trust
orchestrator is invoked.  Upon return from the command handler
the eof status of the socket is inspected and if the socket was
closed while being read, the socket object is reset and poll
command is reconfigured to only poll on two file descriptors.