Properly handle closure of quixote-us control socket.
A regression had been noted with the userspace trust orchestrator
where the trust orchestrator launch process would terminate when
output being generated to the UNIX domain control socket was
interrupted.

The following test is an example of how the regression can be
triggeded.

Shell 1:
quixote-us -c CONTAINER

Shell 2:
quixote-console -T | head -1

The first line of the trajectory output would be properly
generated and the quixote-us orchestrator process would terminate
leaving the namespace monitor process running along with the
workload process.

This was caused by the LocalDuct object not detecting that the
receiving end of the socket had closed the socket connection.

A fix for this was implemented in the LocalDuct object in the
following commit:

9f66a6fee201 ("Properly handle command socket closures.")

The _send_output() private function was implemented in order to
handle the sending of a buffer over the control socket with
detection for the receiving end closing the socket early.

This function sends the buffer over the socket and signals
whether or not an error occurred by setting the return value
through a pointer to the return value supplied by the called.
The caller returns an error condition in its return value but
indicates through direct modification of the return value in the
caller as to whether or not this is a fatal error or the closure
of the socket.

The error condition in the return value causes the caller to
invoke its error handling response but the error returned to its
caller will have been set by the _send_output() function.

Update all of the callers in the trust orchestration process to
use this convention.