Handle closures of control socket in kernel orchestrator.
The quixote orchestrator for in kernel modeling had the same
issue that the userspace orchestrator had.  Closing the receiving
end of the control socket caused termination of the orchestrator
process.

This issue was addressed by introducing the _send_output()
function which is identical to the function that was implemented
for orchestrators that are using userspace modeling.  The
function checks for socket termination and returns an error code
as its return value but sets the error return value in the sender
to a true value to indicate that the orchestrator was not in an
error condition state.

Adjust all of the previous calls to the mgmt->send_Buffer method
to use a call to _sent_output instead.

The child_monitor() function was updated to follow the same
strategy that was implemented in the TSEMworkload object.  An
additional array member was added to the array of pollfd
structures in order to allow the file descriptor from the console
to be polled as well as the file descriptor for the socket
itself.

After a request to the management socket is received the poll
function is configured to poll over both file descriptors for a
change in input to the socket.

After the management command is processed the EOF status is
checked and monitoring of the file descriptor on which commands
are being received is terminated if the remote client has closed
the socket.